requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6
python-dotenv==1.0.0
aiofiles==23.2.1

# Database
firebase-admin==6.4.0


# External APIs
google-cloud-vision==3.5.0
requests==2.31.0
pillow==10.1.0

# Background Jobs
apscheduler==3.10.4

# Authentication & Security
python-jose==3.3.0
bcrypt==4.1.1
cryptography==41.0.7

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Utilities
python-magic==0.4.27
